% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmgpt.R
\name{FMGPT}
\alias{FMGPT}
\title{FM-GPT Fine Mapping Function}
\usage{
FMGPT(
  X,
  Y,
  A,
  group,
  response_types,
  epsilon,
  B,
  burnin,
  thin = 1,
  k_fix = NULL,
  s2.fit = TRUE,
  Sigma.fit = TRUE,
  scale.eta = FALSE
)
}
\arguments{
\item{X}{Covariate matrix X}

\item{Y}{Output matrix Y, possible containing mixed data types}

\item{A}{Indicator for eQTL status of a SNP}

\item{group}{Grouping indicator}

\item{epsilon}{A numeric giving the cutoff for loadings before a factor is dropped}

\item{B}{The total number of iterations}

\item{burnin}{The total number of burnin iterations}

\item{thin}{The thinning parameter}

\item{k_fix}{Fixed number of factors to use}

\item{s2.fit}{Logical indicating whether s^2 should be estimated}

\item{scale.eta}{A logical indicating whether the factor scores should be mean centered}

\item{respones_types}{A vector indicating the variable types in Y}

\item{Sigma}{Logical indicating whether Sigma should be estimated}
}
\value{
A list containing the kept iterations for the loading matrix, number of factors,
beta, a1 and a2, sigma2, sigma_beta, count data parameter r, adjustment effect for eQTL SNPs
}
\description{
FM-GPT Fine Mapping Function
}
